requirements.txt
beautifulsoup4==4.11.1
Flask==2.1.2
flask_session==0.4.0
python_dateutil==2.8.2
requests==2.28.1
lxml
